feat(networking): Configure hostname via module argument
Introduces the `host` argument to the networking module to allow setting
the `networking.hostName` option.  This centralizes hostname
configuration
and ensures consistency.  It relies on the `host` variable being passed
from the host configuration.

This avoids having to specify the hostname within the module itself and
allows it to be managed at a higher level (e.g., host configuration).