fix(system): Correctly pass hostname to system auto-update
Pass the `host` to `system.autoUpdate` for host specific configurations.

This resolves an issue where the system auto-update was failing due to
the
hostname not being correctly configured.  By passing the `host` value as
an argument, the auto-update process can now use the correct hostname.
Relies on
hostname being made available to modules via the hosts configuration.